Acrylic Tunnel Model
Since the first level is underwater, I thought it would be cool for the witch to travel through a tunnel and see outside elements like fish or rocks. This is just a basic tunnel for now. We could later modify it to have a dimensional or magical appearance. Each level might have a different tunnel design, or some levels might not require a closed tunnel at all, especially if they're set outdoors.