Add Potion Bomb Model
I modeled this potion bomb in Blender. We can reuse it for different types of potion bombs, such as red, purple, or yellow ones. I also disabled the particles in the potion bomb model because they were interfering with the glass transparency. We could tweak the particles later so they don't affect the glass.